adjusted delete of model to also delete all surveys of the deleted users

+++ src/entities/user/user.model.js
@@ -1,5 +1,6 @@
 const userSchema = require('./user.schema')
 const dbLoader = require('../../utils/dbLoader')
+const surveyModel = require('../survey/survey.model')()
 
 module.exports = () => {
   const User = dbLoader.getDB().model('user', userSchema, 'user')
@@ -10,7 +11,7 @@ module.exports = () => {
     get: async (find, limit, offset, sort) => {
       try {
         const users = await User.find(find).limit(limit).skip(offset).sort(sort)
-        if (users.length === 0) throw new Error('No user found.')
+        if (users.length === 0) throw new Error('No User found.')
         return users
       } catch (e) {
         throw e
@@ -28,7 +29,7 @@ module.exports = () => {
       try {
         if (Object.prototype.hasOwnProperty.call(data, 'email') && !(await isEmailFree(data.email))) throw new Error('Email already in use. Could not update user.')
         const result = await User.updateMany(where, data)
-        if (result.nMatched === 0) throw new Error('User not found.')
+        if (result.nMatched === 0) throw new Error('No User found.')
         if (result.nModified === 0) throw new Error('User update failed.')
         const updatedUsers = await User.find(where)
         return updatedUsers
@@ -38,12 +39,23 @@ module.exports = () => {
     },
     delete: async (where) => {
       try {
+        const deletedUsers = await User.find(where)
+        if (deletedUsers.length === 0) throw new Error('No User found.')
         const result = await User.deleteMany(where)
         if (result.n === 0) throw new Error('User deletion failed.')
-        // TODO:
-        // Delete Surveys
-        // Delete Contexts if no other user
-        // Delete device reference to user and device if no user
+        /** Delete Surveys of this user * */
+        try {
+          const userIds = deletedUsers.reduce((acc, user) => [...acc, user.id], [])
+          await surveyModel.delete({ creator: { $in: userIds } })
+        } catch (e) {
+          // TODO:
+          // ggf. Modul erstellen, welches fehlgeschlagene DB-Zugriffe
+          // in bestimmten abständen wiederholt
+          // (nur für welche, die nicht ausschlaggebend für erfolg der query sind)
+          console.log(e)
+        }
+        //  TODO: Delete Contexts if no other user
+        //  TODO: Delete device reference to user and device if no user
         return result
       } catch (e) {
         throw e
